reach 12bit pwm

+++ cct_new/manual_ctrl.h
@@ -2,14 +2,12 @@
 #ifndef __MANUAL_CTRL_H
 #define __MANUAL_CTRL_H
 
-
-
 #include <stdint.h> 
 #include <stdbool.h>
 
 #define LED_PWM_CHANNEL_NUM  6
 struct led_pwm_s {
-	uint8_t  pwm_value[LED_PWM_CHANNEL_NUM];
+	uint16_t  pwm_value[LED_PWM_CHANNEL_NUM];
 	uint8_t  pwm_freq ;
 	uint8_t  sync_dimmer;
 	uint16_t dmx_addr ;
